refactor: stores -> store
